Added a request.auth check to logout
In case of no authentication is passed it results a fake successful message in logout. Added a check to request.auth on top of logout

+++ dj_rest_auth/views.py
@@ -150,6 +150,11 @@ class LogoutView(APIView):
         return self.logout(request)
 
     def logout(self, request):
+        if not request.auth:
+            return Response(
+                {'detail': _('You should be logged in to logout. Check whether the token is passed.')},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
         try:
             request.user.auth_token.delete()
         except (AttributeError, ObjectDoesNotExist):
